Starlight-Engine-Mark-1-main/_archive_mark1/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "starlight"
version = "0.3.0"
description = "A high-performance hybrid game engine."
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "numpy>=2.0.0",
    "pillow>=10.3.0",
    "loguru>=0.7.0",
    "pydantic>=2.0.0",
    "scipy>=1.10.0"
]

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "starlight.backend"
python-source = "pysrc"
code-coverage = true

[tool.ruff]
line-length = 120
target-version = "py310"
